+++ HealthCare/src/healthcare/User.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package healthcare;
 
 import java.util.ArrayList;
@@ -15,18 +10,33 @@ public class User {
 
     private String firstName;
     private String lastName;
-    private long userId;
+    private int userId;
     private String password;
-    private long permissions;
+    private int permissions;
     private String dob;
     private String address;
-    private long phoneNumber;
-    private long SSN;
+    private int phoneNumber;
+    private int SSN;
     private String healthInsurance;
     private ArrayList<AppointmentInformation> appointmentInformation = new ArrayList();
     private ArrayList<PaymentInformation> paymentInformation = new ArrayList();
 
-    public User(String firstName, String lastName, long userId, String password, long permissions, String dob, String address, long phoneNumber, long SSN, String healthInsurance) {
+    /**
+     * Param Constructor
+     *
+     * @param firstName First Name of the user
+     * @param lastName Last Name of the user
+     * @param userId User's ID
+     * @param password User's password
+     * @param permissions User's permissions, any numbers can be picked here.
+     * This will tell if it's an staff/doctor/patient
+     * @param dob User's Date of Birth
+     * @param address User's Address
+     * @param phoneNumber User's Phone number
+     * @param SSN User's SSN
+     * @param healthInsurance User's Health insurance provider
+     */
+    public User(String firstName, String lastName, int userId, String password, int permissions, String dob, String address, int phoneNumber, int SSN, String healthInsurance) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.userId = userId;
@@ -39,101 +49,224 @@ public class User {
         this.healthInsurance = healthInsurance;
     }
 
+    /**
+     * Empty constructor
+     */
     public User() {
     }
 
+    /**
+     * Get First name
+     *
+     * @return Returns the first name
+     */
     public String getFirstName() {
         return firstName;
     }
 
+    /**
+     * Set First name
+     *
+     * @param firstName New First name for user
+     */
     public void setFirstName(String firstName) {
         this.firstName = firstName;
     }
 
+    /**
+     * Get Last name
+     *
+     * @return Returns the last name
+     */
     public String getLastName() {
         return lastName;
     }
 
+    /**
+     * Get Appointment Information
+     *
+     * @return An ArrayList of Appointments
+     */
     public ArrayList<AppointmentInformation> getAppointmentInformation() {
         return appointmentInformation;
     }
 
+    /**
+     * Set Appointment Information
+     *
+     * @param appointmentInformation An ArrayList of appointments
+     */
     public void setAppointmentInformation(ArrayList<AppointmentInformation> appointmentInformation) {
         this.appointmentInformation = appointmentInformation;
     }
 
+    /**
+     * Get Payment Information
+     *
+     * @return An ArrayList of payment informations
+     */
     public ArrayList<PaymentInformation> getPaymentInformation() {
         return paymentInformation;
     }
 
+    /**
+     * Set Payment Information
+     *
+     * @param paymentInformation An ArrayList of payment informations
+     */
     public void setPaymentInformation(ArrayList<PaymentInformation> paymentInformation) {
         this.paymentInformation = paymentInformation;
     }
 
+    /**
+     * Set Last name
+     *
+     * @param lastName New last name
+     */
     public void setLastName(String lastName) {
         this.lastName = lastName;
     }
 
-    public long getUserId() {
+    /**
+     * Get User ID
+     *
+     * @return Will return the User's ID
+     */
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    /**
+     * Set User ID
+     *
+     * @param userId New User ID
+     */
+    public void setUserId(int userId) {
         this.userId = userId;
     }
 
+    /**
+     * Get Password
+     *
+     * @return The User's password
+     */
     public String getPassword() {
         return password;
     }
 
+    /**
+     * Set Password
+     *
+     * @param password A new Password for the user
+     */
     public void setPassword(String password) {
         this.password = password;
     }
 
+    /**
+     * Get Permissions
+     *
+     * @return The User's Permissions
+     */
     public long getPermissions() {
         return permissions;
     }
 
-    public void setPermissions(long permissions) {
+    /**
+     * Set Permissions
+     *
+     * @param permissions Set's a new persmission for the user
+     */
+    public void setPermissions(int permissions) {
         this.permissions = permissions;
     }
 
+    /**
+     * Get DOB
+     *
+     * @return Will return the Date of Birth in a mm/dd/yyyy string format
+     */
     public String getDob() {
         return dob;
     }
 
+    /**
+     * Set DOB
+     *
+     * @param dob Will set the Date of Birth, in mm/dd/yyyy format
+     */
     public void setDob(String dob) {
         this.dob = dob;
     }
 
+    /**
+     * Get Address
+     *
+     * @return The User's address
+     */
     public String getAddress() {
         return address;
     }
 
+    /**
+     * Set Address
+     *
+     * @param address New User address
+     */
     public void setAddress(String address) {
         this.address = address;
     }
 
-    public long getPhoneNumber() {
+    /**
+     * Get Phone Number
+     *
+     * @return The User's phone number
+     */
+    public int getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(long phoneNumber) {
+    /**
+     * Set Phone number
+     *
+     * @param phoneNumber Set a new phone number
+     */
+    public void setPhoneNumber(int phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
 
-    public long getSSN() {
+    /**
+     * Get SNN
+     *
+     * @return The User's SSN
+     */
+    public int getSSN() {
         return SSN;
     }
 
-    public void setSSN(long snn) {
+    /**
+     * Set SNN
+     *
+     * @param snn New SSN for User
+     */
+    public void setSSN(int snn) {
         this.SSN = SSN;
     }
 
+    /**
+     * Get Health Insurance
+     *
+     * @return The User's Health Insurance
+     */
     public String getHealthInsurance() {
         return healthInsurance;
     }
 
+    /**
+     * Set Health Insurance
+     *
+     * @param healthInsurance New Health Insurance for User
+     */
     public void setHealthInsurance(String healthInsurance) {
         this.healthInsurance = healthInsurance;
     }
